Factored out the GWTP hosting logic into a new base presenter class
The new base class now includes the final piece of the puzzle: it subscribes to the
ResetPresentersEvent and reacts to it by calling internalReset() on the hosted GWTP
presenter. Most of the HAL presenters rely on this event as a trigger for retrieving
their data, and they weren't getting it because we had disconnected them from GWTP's
RootPresenter.